refactor(fraud): Hide inner client in fraud.Client
The `fraud.Client` struct was refactored to hide the `appleapi.Client`
by making it a private `inner` field.

This change improves encapsulation and makes the API of `fraud.Client`
clearer.

Test-only accessor methods were added to `client_internal_test.go` to
allow testing of the internal state of the client.